Remove platform registration API footgun.
Turns out, you can just pass this thing an architecture that is
*completely* different from the one in the platform you are also passing
in and it'll just happily do the wrong thing. While that's pretty wild,
I'm in favor of not having to make this particular mistake again the
next time I copy/paste some code, so we're now deprecating it.